+++ app/Http/Controllers/NewBorrowingController.php
@@ -6,6 +6,7 @@ use App\Borrowing;
 use App\Http\Requests\NewBorrowingRequest;
 use App\InventoryItem;
 use App\InventoryItemStatus;
+use App\User;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\Auth;
 
@@ -27,10 +28,11 @@ class newBorrowingController extends Controller
         foreach (request('borrowedItems') as $borrowedItem) {
             $startDate = Carbon::createFromFormat('d/m/Y', request('startDate'));
             $expectedReturnDate = Carbon::createFromFormat('d/m/Y', request('expectedReturnDate'));
+            $borrowerId = User::where('email', request('borrowerEmail'))->select('id')->first()->id;
             Borrowing::create([
                 'inventory_item_id' => $borrowedItem,
-                'borrower_id' => 1, // to change when authentication will be setup
-                'initial_lender_id' => Auth::user()->id,  // to change when authentication will be setup
+                'borrower_id' => $borrowerId,
+                'initial_lender_id' => Auth::user()->id,
                 'guarantee' => request('guarantee'),
                 'finished' => false,
                 'start_date' => $startDate,