+++ app/User.php
@@ -2,11 +2,8 @@
 
 namespace App;
 
-use Illuminate\Auth\MustVerifyEmail;
-use Illuminate\Notifications\Notifiable;
 use Illuminate\Foundation\Auth\User as Authenticatable;
-use Illuminate\Contracts\Auth\MustVerifyEmail as MustVerifyEmailContract;
-use Illuminate\Support\Facades\App;
+use Illuminate\Notifications\Notifiable;
 
 class User extends Authenticatable
 {